pyproject.toml
[project]
name = "openrouter-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
